pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jqueue"
version = "0.1.0"
description = "Storage-agnostic job queue using object storage with compare-and-set semantics"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
s3 = ["aioboto3>=13.0"]
gcs = ["google-cloud-storage>=2.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "aioboto3>=13.0",
    "google-cloud-storage>=2.0",
    "ruff",
    "mypy",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
strict = true
python_version = "3.12"
